Split PDF import file handlers
