+++ scripts/quality-baseline.json
@@ -7,6 +7,6 @@
   "viewRuntimeBridgeLookups": 0,
   "labStateAppFiles": 0,
   "labStateTestFiles": 0,
-  "largeJsFilesOver800Lines": 1,
-  "maxJsFileLines": 1026
+  "largeJsFilesOver800Lines": 0,
+  "maxJsFileLines": 799
 }
