+++ service-worker.js
@@ -203,6 +203,7 @@ const APP_SHELL = [
   '/js/image-utils.js',
   '/js/pdf-import.js',
   '/js/pdf-import-commit.js',
+  '/js/pdf-import-file-handlers.js',
   '/js/pdf-import-file-utils.js',
   '/js/pdf-import-spreadsheet.js',
   '/js/pdf-import-preflight.js',
