+++ tests/test-correctness-phase2.js
@@ -115,6 +115,7 @@ const pwaAppShellAssets = [
   '/js/import-drop-zone.js',
   '/js/pdf-import.js',
   '/js/pdf-import-commit.js',
+  '/js/pdf-import-file-handlers.js',
   '/js/pdf-import-file-utils.js',
   '/js/pdf-import-spreadsheet.js',
   '/js/pdf-import-preflight.js',
