+++ tsconfig.json
@@ -228,6 +228,7 @@
     "js/onboarding-view-runtime.js",
     "js/onboarding-view.js",
     "js/pdf-import-commit.js",
+    "js/pdf-import-file-handlers.js",
     "js/pdf-import.js",
     "js/pdf-import-marker-mapping.js",
     "js/pdf-import-marker-normalization.js",
